+++ CHANGELOG.md
@@ -12,3 +12,4 @@ Version 0.4
 * Replaced Pylint W0312 with a custom checker. This means that warnings are only generated for inconsistent indentation characters, rather than warning if spaces were not used.
 * Some messages will now be combined if Pylint generates multiple warnings per line for what is the same cause. For example, 'unused import from wildcard import' messages are now combined rather than having one message per unused import from that line.
 * Messages from multiple tools will be merged if they represent the same problem.
+* Tool failure no longer kills the prospector process but adds a message instead.